+++ openharmony/arkui_ace_engine/frameworks/core/components_ng/pattern/grid/grid_layout_info.h
@@ -371,8 +371,8 @@ struct GridLayoutInfo {
 
     Axis axis_ = Axis::VERTICAL;
 
-    float currentOffset_ = 0.0f; // offset on the current top GridItem on [startMainLineIndex_]
-    float prevOffset_ = 0.0f;
+    double currentOffset_ = 0.0; // offset on the current top GridItem on [startMainLineIndex_]
+    double prevOffset_ = 0.0;
     float currentHeight_ = 0.0f; // height from first item to current top GridItem on [startMainLineIndex_]
     float prevHeight_ = 0.0f;
     float lastMainSize_ = 0.0f;
@@ -394,7 +394,7 @@ struct GridLayoutInfo {
 
     int32_t jumpIndex_ = EMPTY_JUMP_INDEX;
     int32_t jumpForRecompose_ = EMPTY_JUMP_INDEX; // new mark index to notify frontend recomposition
-    std::optional<float> extraOffset_;
+    std::optional<double> extraOffset_;
     int32_t crossCount_ = 0;
     int32_t childrenCount_ = 0;
     int32_t firstRepeatCount_ = 0;
