+++ openharmony/arkui_ace_engine/frameworks/core/components_ng/pattern/waterflow/water_flow_pattern.h
@@ -265,7 +265,7 @@ private:
     RefPtr<WaterFlowLayoutInfoBase> layoutInfo_ = WaterFlowLayoutInfoBase::Create(LayoutMode::TOP_DOWN);
     RefPtr<WaterFlowSections> sections_;
 
-    float prevOffset_ = 0.0f;
+    double prevOffset_ = 0.0;
     SizeF lastSize_;
     std::pair<int32_t, int32_t> itemRange_ = { -1, -1 };
     WeakPtr<UINode> footer_;
